Refactor: 코드 merge 위한 컴포넌트 수정 & 분리
